+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to `astrotomic/iso639` will be documented in this file
 
+## Unreleased
+
+- support custom aliases and localized ICU language names in `name()`
+- add `find()` to resolve names, aliases, and ISO 639 codes
+
 ## 1.0.0 - 2021-05-10
 
 - initial release